+++ bash/feeders/openskynetwork.sh
@@ -42,7 +42,7 @@ if ! grep -q "^deb .*opensky." /etc/apt/sources.list /etc/apt/sources.list.d/*; 
     if [[ ! -d $RECEIVER_BUILD_DIRECTORY/openskynetwork ]]; then
         LogMessage "Creating the OpenSky Network build directory"
         echo ""
-        mkdir -vp $RECEIVER_BUILD_DIRECTORY/openskynetwork
+        mkdir -v $RECEIVER_BUILD_DIRECTORY/openskynetwork 2>&1 | tee -a $RECEIVER_LOG_FILE
         echo ""
     fi
     LogMessage "Entering the OpenSky Network build directory"
@@ -50,8 +50,8 @@ if ! grep -q "^deb .*opensky." /etc/apt/sources.list /etc/apt/sources.list.d/*; 
 
     LogMessage "Downloading and adding the OpenSky Network apt repository GPG key"
     echo ""
-    wget -v -O $RECEIVER_BUILD_DIRECTORY/openskynetwork/opensky.gpg.pub https://opensky-network.org/files/firmware/opensky.gpg.pub
-    wget -q -O - https://opensky-network.org/files/firmware/opensky.gpg.pub | sudo apt-key add -
+    wget -v -O $RECEIVER_BUILD_DIRECTORY/openskynetwork/opensky.gpg.pub https://opensky-network.org/files/firmware/opensky.gpg.pub 2>&1 | tee -a $RECEIVER_LOG_FILE
+    wget -q -O - https://opensky-network.org/files/firmware/opensky.gpg.pub | sudo apt-key add - 2>&1 | tee -a $RECEIVER_LOG_FILE
     echo ""
     LogMessage "Adding the OpenSky Network apt repository"
     sudo bash -c "echo deb https://opensky-network.org/repos/debian opensky custom > /etc/apt/sources.list.d/opensky.list"
@@ -66,7 +66,7 @@ LogHeading "Installing the OpenSky Network feeder package"
 
 LogMessage "Downloading the latest package lists for all enabled repositories and PPAs"
 echo ""
-sudo apt-get update
+sudo apt-get update 2>&1 | tee -a $RECEIVER_LOG_FILE
 echo ""
 LogMessage "Installing the OpenSky Network fedder package using apt"
 CheckPackage opensky-feeder